# HEIGVD-AMT-Gamification/spring-server/src/main/resources/api-spec.yaml
swagger: '2.0'
info:
  title: Gamification API
  description: Move your app forward with the RFPS API.
  version: "1.0.0"
# the domain of the service
host: localhost:8080
...
schemes:
  - http
  - https
# will be prefixed to all paths
basePath: /v1
paths:
...
  /pointScales:
    get:
      description: 
        The PointScales endpoint returns information about point scales belonging to the application whose id is present in the authentication token. The response includes the id, the name, the description, and the coefficient of each point scale. This response is a list of point scale sorted by the id.
      produces:
        - application/json
      tags:
        - PointScales
      parameters:
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. This token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        200:
          description: OK, an array of point scales.
          schema:
            type: array
            items:
              $ref: '#/definitions/PointScaleOutputDTO'
        400:
          description: Bad request, if Authorization header isn't provided or if it's not properly named.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        500:
          description: Internal server error, if database isn't reachable.
          
    post:
      description: Create a new point scale for this application.
      consumes:
        - application/json
      parameters:
        - in: body
          name: badge
          description: PointScale object that needs to be added to the database.
          required: true
          schema:
              $ref: '#/definitions/PointScaleInputDTO'
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        201:
          description: Created, returns the URL to access the new point scale.
          schema:
            $ref: '#/definitions/LocationPointScale'
        400:
          description: Bad request, if Authorization header isn't provided, if it's not properly named or if json payload is malformed.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        422:
          description: Unprocessable entity,
                       if mandatory fields aren't provided,
                       if mandatory fields provided are empty or contain only spaces,
                       if mandatory fields contain more than 80 characters,
                       if point scale name provided already exists.
        500:
          description: Internal server error, if database isn't reachable.
  /pointScales/{id}:
    get:
      description: Returns the point scale which has the specified id.
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the desired point scale.
          required: true
          type: integer
          format: int64
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        200:
          description: OK
          schema:
            $ref: '#/definitions/PointScaleOutputDTO'
        400:
          description: Bad request, if Authorization header isn't provided or if it's not properly named.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        404:
          description: Not found, if id provided doesn't exist for this application.
        500:
          description: Internal server error, if database isn't reachable.
          
    put:
      description: Modify completely the existing point scale which has the specified id.
      consumes:
        - application/json
      parameters:
        - name: id
          in: path
          description: The id of the point scale to modify.
          required: true
          type: integer
          format: int64
        - name: pointScale
          in: body
          description: PointScale to be updated
          required: true
          schema:
              $ref: '#/definitions/PointScaleInputDTO'
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        204:
          description: No content, if success.
        400:
          description: Bad request, if Authorization header isn't provided, if it's not properly named or if json payload is malformed.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        404:
          description: Not found, if id provided doesn't exist for this application.
        422:
          description: Unprocessable entity,
                       if mandatory fields aren't provided,
                       if mandatory fields provided are empty or contain only spaces,
                       if name provided contains more than 80 characters,
                       if pointScale name provided already exists.
        500:
          description: Internal server error, if database isn't reachable.
          
    delete:
      description: Delete the point scale which has the specified id.
      parameters:
        - name: id
          in: path
          description: The id of the desired point scale.
          required: true
          type: integer
          format: int64
        - name: authenticationToken
          in: header
          description: Token used to authenticate the client application. Token must be preceded by "Bearer " pattern.
          required: true
          type: string
      responses:
        204:
          description: No content, if success.
        400:
          description: Bad request, if Authorization header isn't provided or if it's not properly named.
        401:
          description: Unauthorized, if authentication token provided in the Authorization header is malformed, not signed by the gamification API server or expired. 
        404:
          description: Not found, if id provided doesn't exist for this application.
        500:
          description: Internal server error, if database isn't reachable.
...
definitions:
...
  PointScaleOutputDTO:
    type: object
    required: [pointScale_id, name, description, coefficient]
    properties:
      pointScale_id:
        type: string
        description: Unique identifier representing a specific point scale.
      name:
        type: string
        description: Display name of point scale.
      description:
        type: string
        description: Description of point scale.
      coefficient:
        type: integer
        description: Coeffcient of the point scale.
  PointScaleInputDTO:
    type: object
    required: [name,description, coefficient]
    properties:
      name:
        type: string
        description: Display name of point scale.
      description:
        type: string
        description: Description of point scale.
      coefficient:
        type: integer
        description: Coeffcient of the point scale.
  LocationPointScale:
    type: string 
    required: [locationPointScale]
    properties:
      locationPointScale:
        type: string
        description: Unique URL to access the new object created.
...